Support building NCCL with RDMA-core
